// general/shared/schema.ts
export abstract class Schema {
// ...
    public static get liveSchema(): z.ZodType<BiliTypes.RES.Live.Live> {
        return this.memo("liveSchema", () => z.intersection(this.liveInfoSchema, z.object({
            stream: z.union([z.null(), this.liveStreamSchema]),
            streamExpirationAt: z.number().nullable()
        })))
    }
// ...
}